Created ADSR class, Tremolo class, and VibratoWave class. Edited Note class to add the member variable 'Setting' that holds a string of instrument settings. Removed Piano sound samples from embedded resources and used path access instead. Edited SongScore to test Organ and Piano instruments. Edited AR class, and SineWave class to have both generic and specific constructors and added access functions to certain member variables.

+++ Synthie/AudioNode.cs
@@ -2,7 +2,7 @@
 {
     public abstract class AudioNode
     {
-        protected double[] frame = new double[2];
+        protected float[] frame = new float[2];
         protected double samplePeriod;
         protected int sampleRate;
         protected double bpm;
@@ -11,14 +11,14 @@
         /// Access a generated audio frame
         /// </summary
         /// ><returns>the sound sample</returns>
-        public double[] Frame() { return frame; }
+        public float[] Frame() { return frame; }
 
         /// <summary>
         /// Access one channel of a generated audio frame
         /// </summary>
         /// <param name="c">the channel</param>
         /// <returns>the value of at the channel in the sound sample</returns>
-        public double Frame(int c) { return frame[c]; }
+        public float Frame(int c) { return frame[c]; }
 
         public int SampleRate { get => sampleRate; set => sampleRate = value; }
